+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1-rc6] - 2023-11-21
+
+LineProcessingFuncBase no longer takes a `processor` argument (nor does it proviede `_processor`).
+To migrate your code change `super().__init__(processor)` to `super().__init__(); self._processor=processor` if you need to keep the `processor` argument.
+
+### New
+
+ - Fast linting using Ruff
+
+### Changed
+
+ - LineProcessingFuncBase no longer takes a `processor` argument (nor does it proviede `_processor`)
+
 ## [0.1.1-rc5] - 2023-10-17
 
 LineProcessor.add_processor no longer accepts non async functions. Use `LineProcessor.add_processor(wrap_as_async(lambda x: True))` if synchronous functions need to be used as processors.